move ast propagation to its own class
