+++ python-frontend/src/main/java/org/sonar/python/semantic/v2/TypeInferenceV2.java
@@ -37,6 +37,7 @@ import org.sonar.plugins.python.api.tree.Parameter;
 import org.sonar.plugins.python.api.tree.StatementList;
 import org.sonar.plugins.python.api.tree.Tree;
 import org.sonar.python.semantic.SymbolUtils;
+import org.sonar.python.semantic.v2.types.AstBasedPropagation;
 import org.sonar.python.semantic.v2.types.FlowSensitiveTypeInference;
 import org.sonar.python.semantic.v2.types.Propagation;
 import org.sonar.python.semantic.v2.types.PropagationVisitor;
@@ -125,7 +126,8 @@ public class TypeInferenceV2 {
     statements.accept(tryStatementVisitor);
     if (tryStatementVisitor.hasTryStatement()) {
       // CFG doesn't model precisely try-except statements. Hence we fallback to AST based type inference
-      return propagationVisitor.processPropagations(getTrackedVars(declaredVariables, assignedNames));
+      return new AstBasedPropagation(propagationVisitor.propagationsByLhs(), projectLevelTypeTable)
+        .processPropagations(getTrackedVars(declaredVariables, assignedNames));
     }
 
     ControlFlowGraph cfg = controlFlowGraphSupplier.get();
